Fixed king-pawn check issues.
Wrong behaviour: A king could move itself into check if the attacking
piece was a pawn.
Reason: Passive pawn moves were previously discarded if they did not
result in the immediate capture of a piece.
Resolution: Storing the disallowed but controlling moves of the pawn in
an "attacking" key of moves, and adding that key to the space of passive
enemy moves which is searched when looking up king moves. This
limitation only applies to kings.
Fixed: Yes