Started using pdb breakpoints for debugging.
Rewriting the internal state to stop using piece.position attribute.
Move now sets piece.is_active instead of popping from the move list,
since i may change the architecture of storing previous states. It may
become more convenient to instead just use the same pieces objects in
memory and update their properties, and then store board histories as
a list of "hashes" which will just be the FEN string to enable a nice
lookup system.

Started implementing right to castle checks, need to define a mechanism
in find_king_moves which will check for the right to castle, then define
a default false parameter of `Move`, `is_castle`. This can either be
False, 1 or 2 for kingside or queenside castling. This flag should be
checked by move, and the parser also needs to be able to handle the
designation for theses moves; "O-O" and "O-O-O".