Improved the working of the move function.
The moving function now has a much more clear goal; it will just store
changes in the state by freezing it every iteration instead of trying to
bake copying functons into the architecture of state.

State can now manipulate its own instance through the move method, which
takes a move object and updates pieces accordingly.

This bypasses the need to be able to pass the pieces between objects and
instead means that I can migrate the `construct_board` function in
__main__ into the initialiser for Board instead. In this way we will be
able to simply create a state from a fen string and hash it into  a fen.