Switched over to the new ResultView class
In almost all cases now this is used, much cleaner syntax. Still not
included with the __intersection method of Board, simply because I plan
on removing this too at some point and making its usage a whole lot
cleaner. Also removed subtraction of two positions in favour of a new
method find_path with cleaner usage.