Improved castling ability to parse the O-O and O-O-O moves.
Game now runs roughly as expected including castling. Noticed some bugs
around king behaviour with pawn attacking squares, so redefining the use
of the result.attack set to be every hypothetical move (regardless of if
it is legal) so we do not allow kings to move into checks indirectly, or
mark squares as illegal when they are in fact legal.