Major changes
Changed the Position API to be in sync with the C++ version, since only
one needs to be speed optimised. Fixed all the associated bugs, so now
all the tests except for those which still dont work because of the
board api pass. There are a couple of improvements - the factories need
to respect the USE_CPP setting in constants.py, and there needs to be
a better way to get that setting passed in, rather than hardcoding it.
Additionally, i need to write a factory method for the Board, remove
the Game class and refactor that into a small library of helper
functions rather than a big class.

Updated the version of pychessbinds to include feature parity of the
Piece objects generated from its methods. There is now a complete
ecosystem either in C++ or in Python, dependent on what is available.

The most important thing now is to develop the tests - this is gonna be
pretty much the only way to catch bugs at this point. Tests need to be
run on both versions of the State, Coordinate and Piece
implementations.

Finally, helpers need to be refactored into a proper system outside of
state - the parse_move needs to be moved inside the base Board
implementation, but this is going to be really busy. It might be worth
looking to see if abstract base classes can help with keeping the code
tidier, or if its just worth having one big class.

The board factory function should provide one coherent API at the end.
There should be a lear way to:
- Get the correct version of important classes such as Position, Piece
- Interact with the boards for:
  - Finding piece positions
  - Moving a piece
  - Game state (metadata and winstate)
  - Loading a new game or providing a FEN string for setup (move
    pieces_from_fen to inside the constructor perhaps? Then it can also
    benefit from factories)